Improve secret handling
The isManagedTLSSecret helper relied on ingresses, but wasn't checked on
an update to ingresses. This instead moves all checks to the plugin.

In addition, create a secret informer builds a cache of all secrets
in memory. This is undesirable, because it means we keep around
sensitive data from completely unrelated secrets. It is especially
harmful when using caddy-ingress as the default cluster-wide ingress.

Instead, a transform is installed on the informer to throw away all
secret data, so we only keep metadata. The plugin is then responsible
for picking up changes and fetching actual secret data for only the
secrets we need.

In practice, we are still dependant on the Go garbage collector.